Remove use of camlp4 lwt syntax
I'm unsure about the for_lwt loop, as I don't know whether it's executed
sequentially or in parallel, and using a list instead might use more
memory for a large m.

Also, I assumed that the >> sequencing operator has the same meaning in
the camlp4 syntax extension as the ppx one.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>